Changed all  JS scripts
