+++ public/index.html
@@ -32,13 +32,10 @@
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
-    <script async="async" src="scripts/showMenu.js"></script>
-<!--    <script async="async" src="scripts/changeTheme.js"></script>-->
-    <script async="async" src="scripts/scroll.js"></script>
 
-    <script async="async" src="scripts/qualificationList.js"></script>
-    <script async="async" src="scripts/showServices.js"></script>
-    <script async="async" src="scripts/portfolioSwiper.js"></script>
+<!--    <script async="async" src="scripts/scroll.js"></script>-->
+
+    <script async="async" src="scripts/swiper.js"></script>
     <script src="scripts/swiper-bundle.min.js"></script>
   </body>
 </html>
