+++ src/components/body/services/script.js
@@ -1,0 +1,23 @@
+export const showServices = () => {
+    const modalViews = document.querySelectorAll('.services__modal'),
+        modalBtns = document.querySelectorAll('.services__button'),
+        modalCloses = document.querySelectorAll('.services__modal-close');
+
+    let modal = function (modalClick) {
+        modalViews[modalClick].classList.add('active-modal')
+    }
+
+    modalBtns.forEach((modalBtn, i) => {
+        modalBtn.addEventListener('click', () => {
+            modal(i)
+        })
+    })
+
+    modalCloses.forEach((modalClose) => {
+        modalClose.addEventListener('click', () => {
+            modalViews.forEach((modalView) => {
+                modalView.classList.remove('active-modal')
+            })
+        })
+    })
+}
