+++ src/components/scroll-up/index.js
@@ -1,5 +1,7 @@
 import React from 'react';
 import './scroll.css'
+import {scroll_up} from "./script";
+
 
 function ScrollUp() {
     return (
